fix(test): merge conftest.py files
